ibig::modular
Struct ModuloRing

pub struct ModuloRing(/* private fields */);

A ring of integers modulo a positive integer.

Examples

let ring = ModuloRing::new(&ubig!(100));
assert_eq!(ring.modulus(), ubig!(100));

Implementations

impl ModuloRing

pub fn modulus(&self) -> UBig

The ring modulus.

Example

let ring = ModuloRing::new(&ubig!(100));
assert_eq!(ring.modulus(), ubig!(100));

pub fn from<T: IntoModulo>(&self, x: T) -> Modulo<'_>

Create an element of the ring from another type.

Examples

let ring = ModuloRing::new(&ubig!(100));
let x = ring.from(-1234);
let y = ring.from(ubig!(3366));
assert!(x == y);

impl ModuloRing

pub fn new(n: &UBig) -> ModuloRing

Create a new ring of integers modulo n.

For two Modulo numbers to be compatible, they must come from the same ModuloRing. Two different ModuloRings are not compatible even if they have the same modulus n.

Examples

let ring = ModuloRing::new(&ubig!(100));
assert_eq!(ring.modulus(), ubig!(100));

Panics

Panics if n is zero.