ibig
Struct UBig

pub struct UBig(/* private fields */);

Unsigned big integer.

Arbitrarily large unsigned integer.

Examples

let a = ubig!(a2a123bbb127779cccc123123ccc base 32);
let b = ubig!(0x1231abcd4134);
let c = UBig::from_str_radix("a2a123bbb127779cccc123123ccc", 32)?;
let d = UBig::from_str_radix("1231abcd4134", 16)?;
assert_eq!(a, c);
assert_eq!(b, d);

Implementations

impl UBig

pub fn bit(&self, n: usize) -> bool

Returns true if the n-th bit is set.

Examples

assert_eq!(ubig!(0b10010).bit(1), true);
assert_eq!(ubig!(0b10010).bit(3), false);
assert_eq!(ubig!(0b10010).bit(100), false);

pub fn set_bit(&mut self, n: usize)

Set the n-th bit.

Examples

let mut a = ubig!(0b100);
a.set_bit(0);
assert_eq!(a, ubig!(0b101));
a.set_bit(10);
assert_eq!(a, ubig!(0b10000000101));

pub fn clear_bit(&mut self, n: usize)

Clear the n-th bit.

Examples

let mut a = ubig!(0b101);
a.clear_bit(0);
assert_eq!(a, ubig!(0b100));

pub fn trailing_zeros(&self) -> Option<usize>

Returns the number of trailing zeros in the binary representation.

In other words, it is the largest n such that 2 to the power of n divides the number.

For 0, it returns None.

Examples

assert_eq!(ubig!(17).trailing_zeros(), Some(0));
assert_eq!(ubig!(48).trailing_zeros(), Some(4));
assert_eq!(ubig!(0b101000000).trailing_zeros(), Some(6));
assert_eq!(ubig!(0).trailing_zeros(), None);

pub fn bit_len(&self) -> usize

Bit length.

The length of the binary representation of the number.

For 0, the length is 0.

For non-zero numbers it is:

in_radix(2).to_string().len()

the index of the top 1 bit plus one

the floor of the logarithm base 2 of the number plus one.

Examples

assert_eq!(ubig!(17).bit_len(), 5);
assert_eq!(ubig!(0b101000000).bit_len(), 9);
assert_eq!(ubig!(0).bit_len(), 0);
let x = ubig!(_0x90ffff3450897234);
assert_eq!(x.bit_len(), x.in_radix(2).to_string().len());

pub fn is_power_of_two(&self) -> bool

True if the number is a power of 2.

Examples

assert_eq!(ubig!(0).is_power_of_two(), false);
assert_eq!(ubig!(8).is_power_of_two(), true);
assert_eq!(ubig!(9).is_power_of_two(), false);

impl UBig

pub fn from_le_bytes(bytes: &[u8]) -> UBig

Construct from little-endian bytes.

Examples

assert_eq!(UBig::from_le_bytes(&[3, 2, 1]), ubig!(0x010203));

pub fn from_be_bytes(bytes: &[u8]) -> UBig

Construct from big-endian bytes.

Examples

assert_eq!(UBig::from_be_bytes(&[1, 2, 3]), ubig!(0x010203));

pub fn to_le_bytes(&self) -> Vec<u8>

Return little-endian bytes.

Examples

assert!(ubig!(0).to_le_bytes().is_empty());
assert_eq!(ubig!(0x010203).to_le_bytes(), [3, 2, 1]);

pub fn to_be_bytes(&self) -> Vec<u8>

Return big-endian bytes.

Examples

assert!(ubig!(0).to_be_bytes().is_empty());
assert_eq!(ubig!(0x010203).to_be_bytes(), [1, 2, 3]);

pub fn to_f32(&self) -> f32

Convert to f32.

Round to nearest, breaking ties to even last bit.

Examples

assert_eq!(ubig!(134).to_f32(), 134.0f32);

pub fn to_f64(&self) -> f64

Convert to f64.

Round to nearest, breaking ties to even last bit.

Examples

assert_eq!(ubig!(134).to_f64(), 134.0f64);

impl UBig

pub fn in_radix(&self, radix: u32) -> InRadix<'_>

Representation in a given radix.

Panics

Panics if radix is not between 2 and 36 inclusive.

Examples

assert_eq!(format!("{}", ubig!(83).in_radix(3)), "10002");
assert_eq!(format!("{:+010}", ubig!(35).in_radix(36)), "+00000000z");

impl UBig

pub fn gcd(&self, rhs: &UBig) -> UBig

Greatest common divisor.

Example

assert_eq!(ubig!(12).gcd(&ubig!(18)), ubig!(6));

Panics

ubig!(0).gcd(&ubig!(0)) panics.

pub fn extended_gcd(&self, rhs: &UBig) -> (UBig, IBig, IBig)

Greatest common divisors and the Bézout coefficients.

If a.extended_gcd(&b) == (g, x, y) then:

x * a + y * b == g

abs(x) <= max(b, 1)

abs(y) <= max(a, 1)

Example

let a = ubig!(12);
let b = ubig!(18);
let (g, x, y) = a.extended_gcd(&b);
assert_eq!(&a % &g, ubig!(0));
assert_eq!(&b % &g, ubig!(0));
assert_eq!(&x * IBig::from(&a) + &y * IBig::from(&b), IBig::from(g));
assert!(x.unsigned_abs() <= b);
assert!(y.unsigned_abs() <= a);

Panics

ubig!(0).extended_gcd(&ubig!(0)) panics.

impl UBig

pub fn from_str_radix(src: &str, radix: u32) -> Result<UBig, ParseError>

Convert a string in a given base to UBig.

src may contain an optional + prefix. Digits 10-35 are represented by a-z or A-Z.

Panics

Panics if radix is not between 2 and 36 inclusive.

Examples

assert_eq!(UBig::from_str_radix("+7ab", 32)?, ubig!(7499));

pub fn from_str_with_radix_prefix(src: &str) -> Result<UBig, ParseError>

Convert a string with an optional radix prefix to UBig.

src may contain an optional + after the radix prefix.

Allowed prefixes: 0b for binary, 0o for octal, 0x for hexadecimal.

Examples

assert_eq!(UBig::from_str_with_radix_prefix("+0o17")?, ubig!(0o17));
assert_eq!(UBig::from_str_with_radix_prefix("0x1f")?, ubig!(0x1f));

impl UBig

pub fn pow(&self, exp: usize) -> UBig

Raises self to the power of exp.

Example

assert_eq!(ubig!(3).pow(3), ubig!(27));

impl UBig

pub const MAX_BIT_LEN: usize = 9_223_372_036_854_775_808usize

Maximum length in bits.

UBigs up to this length are supported. Creating a longer number will panic.

This does not guarantee that there is sufficient memory to store numbers up to this length. Memory allocation may fail even for smaller numbers.

The fact that this limit fits in usize guarantees that all bit addressing operations can be performed using usize.

It is typically close to usize::MAX, but the exact value is platform-dependent.